Added a basic turn function that needs to have proper pid tuning

+++ src/main/cpp/commands/DriveTeleop.cpp
@@ -7,7 +7,7 @@ DriveTeleop::DriveTeleop(Drivetrain* drivetrain) : m_drivetrain(drivetrain){
 }
 
 void DriveTeleop::Execute(){
-    float zRot = m_driveJoystick.GetRawAxis(TURN_RIGHT_AXIS) - m_driveJoystick.GetRawAxis(TURN_LEFT_AXIS);
-    zRot = std::max(std::min(zRot, 1.0f), -1.0f);
+    double zRot = m_driveJoystick.GetRawAxis(TURN_RIGHT_AXIS) - m_driveJoystick.GetRawAxis(TURN_LEFT_AXIS);
+    zRot = std::max(std::min(zRot, 1.0), -1.0);
     m_drivetrain->ArcadeDrive(m_driveJoystick.GetRawAxis(MOVE_X_AXIS), zRot);
 }
